Separate asyncState tests
